refac: change database setup file, remove dir of model in operations package and remove local_Exception class
